model_manager/requirements.txt
Flask==2.0.1
requests==2.26.0
psycopg2-binary==2.9.1
Flask-CORS==3.0.10
